Reorganized Entire Project (Please Read)
- Moved documents into Assets folder
- Consolodated all assets into categories
- Please read "Project Layout.pdf" for more details

Please also note that I cannot guarantee that certain aspects of your
prefabs or scenes did not get moved around, so do not fret over anything
misplaced or not functioning as it was before. Chances are, you just
need to tell your prefab where to locate its assets. Nothing was
deleted, only moved.